test: move locked row test doubles to support files
